+++ packages/medusa/src/api/utils/middlewares/products/index.ts
@@ -1,5 +1,7 @@
 export * from "./filter-by-valid-sales-channels"
 export * from "./normalize-data-for-context"
+export * from "./remap-product-cross-module-filters"
+export * from "./remap-product-search-filters"
 export * from "./set-pricing-context"
 export * from "./set-tax-context"
 export * from "./variant-inventory-quantity"