+++ packages/design-system/icons/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @medusajs/icons
 
+## 2.20.1
+
+## 2.20.0
+
 ## 2.19.0
 
 ### Patch Changes