+++ packages/design-system/ui/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @medusajs/ui
 
+## 4.2.3
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @medusajs/icons@2.20.1
+
+## 4.2.2
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @medusajs/icons@2.20.0
+
 ## 4.2.1
 
 ### Patch Changes